LayoutTests/http/tests/storageAccess/request-and-deny-storage-access-cross-origin-sandboxed-iframe-expected.txt

CONFIRM: Do you want to use your  ID on ?
CONSOLE MESSAGE: line 68: PASS document.hasStorageAccess was denied.
Tests that cross-origin iframe storage access is denied if the iframe is sandboxed, has the allow token, but the user opts out.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
